requirements.txt
Flask==3.1.0 
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
python-dotenv==1.0.1
nltk==3.9.1


